refactor: extract agent submodules and add TUI chat state types
